Add weather back into a class and refactor fail statements into single blocks
