+++ lib/airport.rb
@@ -1,11 +1,12 @@
 require_relative 'plane'
+require_relative 'weather'
 class Airport
 
   AIRPORT_CAPACITY = 20
 
   attr_reader :capacity, :planes
 
-  def initialize
+  def initialize(weather: Weather.new)
     @planes = []
     @capacity = AIRPORT_CAPACITY
   end
@@ -13,34 +14,36 @@ class Airport
   public
 
   def land(plane)
-    fail "Bad weather today. Cannot land." if stormy_weather?
-    fail "Sorry. Airport full. Go away." if full?
-    fail "This plane is already landed." if plane.landed?
+    if stormy_weather?
+      fail "Bad weather today. Cannot land."
+    elsif full?
+      fail "Sorry. Airport full. Go away."
+    elsif plane.landed?
+    fail "This plane is already landed."
+    else
     plane.land_plane
     planes << plane
+    end
   end
 
   def take_off(plane)
-    fail "Bad weather today. Cannot take off." if stormy_weather?
-    fail "That plane is not in this airport" unless planes.include? plane
+    if stormy_weather?
+    fail "Bad weather today. Cannot take off."
+    elsif !planes.include? plane
+    fail "That plane is not in this airport"
+    else
     plane.take_off_plane
     planes.delete(plane)
+    end
   end
 
   def set_capacity(number)
     @capacity = number
   end
 
-  def stormy_weather?
-    random_number = rand(10)
-    if random_number >= 8
-      true
-    else
-      false
-    end
-  end
-
   private
+
+  attr_reader :weather
 
   def full?
     planes.count >= capacity
