+++ lib/weather.rb
@@ -1,0 +1,10 @@
+class Weather
+  def stormy_weather?
+    random_number = rand(10)
+    if random_number >= 8
+      true
+    else
+      false
+    end
+  end
+end
